Reading JSON FILES
This is the method I came up with to read JSON objects. I made this in a react framework right now.
This project was bootstrapped with Create React App.

You guys might have to install react to use this. (As I couldn't install the files as a PR for some reason) 

You could run these commands in the vscode terminal to install react and then copy the records.json and whatever changes I made in app.js:
npx create-react-app my-app
cd my-app
npm start