Interactive Homepage
This is an interactive home page for our website. For the time being, this doesn't work with the flask app. But it is fully functional. I believe this is because the flask app takes file links as static objects. Also, If anyone wants to know how I did this they can use this link https://vincentgarreau.com/particles.js/