ErrorPage
The problem that I am having rn is that the app doesn't work from my home IP. I would try to rewrite these files within the app tom morning. But I wrote them exactly as given in the files by Takashi, so it wouldn't be a problem rewriting them(copying and pasting).