Front-end Final(DO NOT MERGE)
This is the main front-end done for the web app. Everything works. I have already committed and pushed the work in the main branch. Making a PR was making it a bit complicated. So, these are all the files I worked with. We do not need to merge these files with the main one!